Change index math back from bytes to ints. Refactor BFS and DFS better, begin Selective algorithm

+++ math-square/mathsquare/src/main/java/com/mathsquare/solvers/Calculators/IPredictiveCalculator.java
@@ -1,5 +1,5 @@
 package com.mathsquare.solvers.Calculators;
 
 public interface IPredictiveCalculator {
-    void loadOptions(long options, byte justPlaced);
+    void loadOptions(long options, int justPlaced);
 }
